+++ crates/teloxide-core/src/payloads/answer_pre_checkout_query.rs
@@ -2,7 +2,7 @@
 
 use serde::Serialize;
 
-use crate::types::True;
+use crate::types::{PreCheckoutQueryId, True};
 
 impl_payload! {
     /// Once the user has confirmed their payment and shipping details, the Bot API sends the final confirmation in the form of an [`Update`] with the field pre\_checkout\_query. Use this method to respond to such pre-checkout queries. On success, True is returned. **Note:** The Bot API must receive an answer within 10 seconds after the pre-checkout query was sent.
@@ -12,7 +12,7 @@ impl_payload! {
     pub AnswerPreCheckoutQuery (AnswerPreCheckoutQuerySetters) => True {
         required {
             /// Unique identifier for the query to be answered
-            pub pre_checkout_query_id: String [into],
+            pub pre_checkout_query_id: PreCheckoutQueryId,
             /// Specify True if everything is alright (goods are available, etc.) and the bot is ready to proceed with the order. Use False if there are any problems.
             pub ok: bool,
         }