+++ crates/teloxide-core/src/types/file.rs
@@ -1,5 +1,6 @@
 use std::ops::Deref;
 
+use derive_more::derive::From;
 use serde::{Deserialize, Serialize};
 
 /// This object represents a file ready to be downloaded.
@@ -29,18 +30,54 @@ pub struct File {
     pub path: String,
 }
 
+/// Identifier for a file
+#[derive(
+    Default,
+    Clone,
+    Debug,
+    derive_more::Display,
+    PartialEq,
+    Eq,
+    Hash,
+    Serialize,
+    Deserialize,
+    From
+)]
+#[serde(transparent)]
+#[from(&'static str, String)]
+pub struct FileId(pub String);
+
+/// Unique identifier for a file, which is supposed to be the same over
+/// time and for different bots. Can't be used to download or reuse the
+/// file.
+#[derive(
+    Default,
+    Clone,
+    Debug,
+    derive_more::Display,
+    PartialEq,
+    Eq,
+    Hash,
+    Serialize,
+    Deserialize,
+    From
+)]
+#[serde(transparent)]
+#[from(&'static str, String)]
+pub struct FileUniqueId(pub String);
+
 /// Metadata of a [`File`].
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct FileMeta {
     /// Identifier for this file.
     #[serde(rename = "file_id")]
-    pub id: String,
+    pub id: FileId,
 
     /// Unique identifier for this file, which is supposed to be the same over
     /// time and for different bots. Can't be used to download or reuse the
     /// file.
     #[serde(rename = "file_unique_id")]
-    pub unique_id: String,
+    pub unique_id: FileUniqueId,
 
     /// File size in bytes.
     #[serde(rename = "file_size")]
@@ -60,13 +97,13 @@ pub(crate) const fn file_size_fallback() -> u32 {
 /// ## Examples
 ///
 /// ```rust
-/// use teloxide_core::types::File;
+/// use teloxide_core::types::{File, FileId, FileUniqueId};
 /// #
-/// # let get_file = || File { meta: teloxide_core::types::FileMeta { id: String::new(), unique_id: String::new(), size: 0 }, path: String::new() };
+/// # let get_file = || File { meta: teloxide_core::types::FileMeta { id: FileId(String::new()), unique_id: FileUniqueId(String::new()), size: 0 }, path: String::new() };
 /// let file: File = get_file();
 ///
-/// let file_id: &str = &file.id;
-/// let file_unique_id: &str = &file.unique_id;
+/// let file_id: &FileId = &file.id;
+/// let file_unique_id: &FileUniqueId = &file.unique_id;
 /// let file_size: u32 = file.size;
 /// #
 /// # let _ = (file_id, file_unique_id, file_size);
@@ -81,7 +118,7 @@ impl Deref for File {
 
 #[cfg(test)]
 mod tests {
-    use crate::types::{File, FileMeta};
+    use crate::types::{File, FileId, FileMeta, FileUniqueId};
 
     // As per <https://github.com/tdlib/telegram-bot-api/issues/192> file size is **not** optional,
     // But <https://github.com/tdlib/telegram-bot-api/issues/294> suggests that it can be missing in case Telegram servers are going insane.
@@ -96,8 +133,8 @@ mod tests {
             file,
             File {
                 meta: FileMeta {
-                    id: "FILE_ID".to_owned(),
-                    unique_id: "FILE_UNIQUE_ID".to_owned(),
+                    id: FileId("FILE_ID".to_owned()),
+                    unique_id: FileUniqueId("FILE_UNIQUE_ID".to_owned()),
                     size: u32::MAX,
                 },
                 path: "FILE_PATH".to_owned(),
@@ -114,7 +151,11 @@ mod tests {
 
         assert_eq!(
             file,
-            FileMeta { id: "FILE_ID".to_owned(), unique_id: "FILE_UNIQUE_ID".to_owned(), size: 42 }
+            FileMeta {
+                id: FileId("FILE_ID".to_owned()),
+                unique_id: FileUniqueId("FILE_UNIQUE_ID".to_owned()),
+                size: 42
+            }
         );
     }
 
@@ -127,8 +168,8 @@ mod tests {
             file,
             File {
                 meta: FileMeta {
-                    id: "FILE_ID".to_owned(),
-                    unique_id: "FILE_UNIQUE_ID".to_owned(),
+                    id: FileId("FILE_ID".to_owned()),
+                    unique_id: FileUniqueId("FILE_UNIQUE_ID".to_owned()),
                     size: 42,
                 },
                 path: "FILE_PATH".to_owned(),