+++ crates/teloxide-core/src/types/poll.rs
@@ -1,7 +1,25 @@
 use crate::types::{MessageEntity, PollType, Seconds, User};
 
 use chrono::{DateTime, Utc};
+use derive_more::derive::From;
 use serde::{Deserialize, Serialize};
+
+/// Unique poll identifier.
+#[derive(
+    Default,
+    Clone,
+    Debug,
+    derive_more::Display,
+    PartialEq,
+    Eq,
+    Hash,
+    Serialize,
+    Deserialize,
+    From
+)]
+#[serde(transparent)]
+#[from(&'static str, String)]
+pub struct PollId(pub String);
 
 /// This object contains information about a poll.
 ///
@@ -10,7 +28,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct Poll {
     /// Unique poll identifier.
-    pub id: String,
+    pub id: PollId,
 
     /// Poll question, 1-300 characters.
     pub question: String,